Run cloud test command and cleanup (#2)
